requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
django-jsonfield==1.4.0
django-log-formatter-json==0.2.0
django-log-request-id==1.6.0
django-redis==4.10.0
django-requestlogs==0.2.1
djangorestframework==3.11.0
django-environ==0.4.5